Move resources under tfimport
