+++ internal/tfimport/tfimport.go
@@ -21,9 +21,9 @@ import (
 	"os/exec"
 	"regexp"
 
-	"github.com/GoogleCloudPlatform/healthcare-data-protection-suite/internal/resources"
 	"github.com/GoogleCloudPlatform/healthcare-data-protection-suite/internal/runner"
 	"github.com/GoogleCloudPlatform/healthcare-data-protection-suite/internal/terraform"
+	"github.com/GoogleCloudPlatform/healthcare-data-protection-suite/internal/tfimport/resources"
 )
 
 // Regexes used in parsing the output of the `terraform import` command.
