+++ internal/tfimport/tfimport_test.go
@@ -22,8 +22,8 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/GoogleCloudPlatform/healthcare-data-protection-suite/internal/resources"
 	"github.com/GoogleCloudPlatform/healthcare-data-protection-suite/internal/terraform"
+	"github.com/GoogleCloudPlatform/healthcare-data-protection-suite/internal/tfimport/resources"
 	"github.com/google/go-cmp/cmp"
 )
